+++ pypgx/cli/estimate_phase_beagle.py
@@ -10,12 +10,10 @@ description = f"""
 # Estimate haplotype phase of observed variants with the Beagle program. #
 ##########################################################################
 
-If your input data is GRCh37, I recommend using the 1000 Genomes Project phase 3 reference panel. You can easily download it thanks to the authors of Beagle:
 
-$ wget -r --no-parent http://bochet.gcc.biostat.washington.edu/beagle/1000_Genomes_phase3_v5a/b37.vcf/
 
 Usage examples:
-  $ pypgx {fuc.api.common._script_name()} imported-variants.zip ref.vcf phased-variants.zip
+  $ pypgx {fuc.api.common._script_name()} imported-variants.zip phased-variants.zip
 """
 
 def create_parser(subparsers):
@@ -31,13 +29,14 @@ def create_parser(subparsers):
         help='Archive file with the semantic type VcfFrame[Imported].'
     )
     parser.add_argument(
-        'panel',
-        help='Reference haplotype panel.'
-    )
-    parser.add_argument(
         'phased_variants',
         metavar='phased-variants',
         help='Archive file with the semantic type VcfFrame[Phased].'
+    )
+    parser.add_argument(
+        '--panel',
+        metavar='PATH',
+        help='Reference haplotype panel. By default, the 1KGP panel is used.'
     )
     parser.add_argument(
         '--impute',