+++ pypgx/cli/import_variants.py
@@ -6,19 +6,19 @@ import fuc
 import pysam
 
 description = f"""
-########################################
-# Import variant data for target gene. #
-########################################
+############################################
+# Import variant data for the target gene. #
+############################################
 
 Usage examples:
-  $ pypgx {fuc.api.common._script_name()} CYP2D6 variants.vcf CYP2D6-variants-imported.zip
+  $ pypgx {fuc.api.common._script_name()} CYP2D6 input.vcf CYP2D6-imported-variants.zip
 """
 
 def create_parser(subparsers):
     parser = fuc.api.common._add_parser(
         subparsers,
         fuc.api.common._script_name(),
-        help='Import variant data for target gene.',
+        help='Import variant data for the target gene.',
         description=description,
     )
     parser.add_argument(
@@ -30,7 +30,8 @@ def create_parser(subparsers):
         help='VCF file (zipped or unzipped).'
     )
     parser.add_argument(
-        'output',
+        'imported_variants',
+        metavar='imported-variants',
         help='Archive file with the semantic type VcfFrame[Imported].'
     )
     parser.add_argument(
@@ -44,4 +45,4 @@ def main(args):
     archive = utils.import_variants(
         args.gene, args.vcf, assembly=args.assembly
     )
-    archive.to_file(args.output)
+    archive.to_file(args.imported_variants)